xtf: Add monitor test class
This class starts alongside the domain a monitor application which opens
an event channel corresponding to that domain and handles the received
requests.
Use the "monitor_args" key to pass test specific arguments to the
monitor application.
The arguments will be added in the test's Makefile using the
TEST-EXTRA-INFO variable.

Signed-off-by: Petre Pircalabu <ppircalabu@bitdefender.com>