xtf: Add emul-unimpl test
Add a new test to verify if XEN can correctly handle the
X86EMUL_UNIMPLEMENTED event.

The XTF DomU test image just executes a instruction not implemented by
the XEN X86 emulator (fstenv) and checks if the execution was
successfull. This instruction will be the first one in a custom .text
section.

In order to instruct XEN to try to emulate that instruction the monitor
application changes the attributes of that specific page in order to
inhibit execution. This will trigger a MEM_ACCESS request that will be
responded by toggling the EMULATE flag.
The emulation will fail, which will trigger an EMUL_UNIMPLEMENTED
request which will be handled by enabling execution on that specific
page (altp2m) and singlestepping that instruction.

The test will be successfull if the instruction can be executed
correctly.

Signed-off-by: Petre Pircalabu <ppircalabu@bitdefender.com>